// frontend/src/app/interfaces/direcciones.ts
// Generated by https://quicktype.io

export interface DirectionsResponse {
  routes:    Route[];
  waypoints: Waypoint[];
  code:      string;
  uuid:      string;
}

export interface Route {
  weight_name: string;
  weight:      number;
  duration:    number;
  distance:    number;
  legs:        Leg[];
  geometry:    Geometry;
}

export interface Geometry {
  coordinates: Array<number[]>;
  type:        Type;
}

export enum Type {
  LineString = "LineString",
}

export interface Leg {
  via_waypoints: any[];
  admins:        Admin[];
  weight:        number;
  duration:      number;
  steps:         Step[];
  distance:      number;
  summary:       string;
}

export interface Admin {
  iso_3166_1_alpha3: string;
  iso_3166_1:        string;
}

export interface Step {
  intersections: Intersection[];
  maneuver:      Maneuver;
  name:          string;
  duration:      number;
  distance:      number;
  driving_side:  DrivingSide;
  weight:        number;
  mode:          Mode;
  geometry:      Geometry;
}

export enum DrivingSide {
  Left = "left",
  Right = "right",
  Straight = "straight",
}

export interface Intersection {
  classes?:           ClassElement[];
  bearings:           number[];
  entry:              boolean[];
  mapbox_streets_v8?: MapboxStreetsV8;
  is_urban?:          boolean;
  admin_index:        number;
  out?:               number;
  geometry_index:     number;
  location:           number[];
  turn_weight?:       number;
  turn_duration?:     number;
  in?:                number;
  stop_sign?:         boolean;
  weight?:            number;
  duration?:          number;
  yield_sign?:        boolean;
  lanes?:             Lane[];
  traffic_signal?:    boolean;
}

export enum ClassElement {
  Restricted = "restricted",
}

export interface Lane {
  indications:       DrivingSide[];
  valid:             boolean;
  active:            boolean;
  valid_indication?: DrivingSide;
}

export interface MapboxStreetsV8 {
  class: MapboxStreetsV8Class;
}

export enum MapboxStreetsV8Class {
  Primary = "primary",
  Secondary = "secondary",
  SecondaryLink = "secondary_link",
  Service = "service",
  Street = "street",
  StreetLimited = "street_limited",
  Tertiary = "tertiary",
  TertiaryLink = "tertiary_link",
}

export interface Maneuver {
  type:           string;
  instruction:    string;
  bearing_after:  number;
  bearing_before: number;
  location:       number[];
  modifier?:      string;
}

export enum Mode {
  Driving = "driving",
}

export interface Waypoint {
  distance: number;
  name:     string;
  location: number[];
}